+++ modules/spring-core/src/test/kotlin/com/esprito/spring/core/providers/java/SpringLineMarkerUsagesFromBeanArrayTest.kt
@@ -1,0 +1,107 @@
+package com.esprito.spring.core.providers.java
+
+import com.esprito.spring.core.SpringCoreClasses
+import com.esprito.spring.core.SpringIcons
+import com.esprito.spring.core.util.SpringGutterTestUtil
+import com.esprito.spring.test.EspritoJavaLightTestCase
+import com.esprito.spring.test.TestLibrary
+import org.intellij.lang.annotations.Language
+import org.jetbrains.kotlin.test.TestMetadata
+
+private const val TEST_DATA_PATH = "providers/linemarkers/beans"
+
+@TestMetadata(TEST_DATA_PATH)
+class SpringLineMarkerUsagesFromBeanArrayTest : EspritoJavaLightTestCase() {
+
+    override fun getTestDataPath(): String = super.getTestDataPath() + TEST_DATA_PATH
+
+    override val libraries: Array<TestLibrary> = arrayOf(TestLibrary.springContext_6_0_7)
+
+    fun testArrayUsageTargetType() {
+        myFixture.copyFileToProject("BeanUsagesClasses.java")
+        myFixture.addClass(
+            """                                    
+            @${SpringCoreClasses.COMPONENT}
+            class TestUsages {
+                @${SpringCoreClasses.AUTOWIRED}
+                B[] b;                
+            }
+        """.trimIndent()
+        )
+
+        @Language("JAVA") val text = """           
+            @${SpringCoreClasses.CONFIGURATION}
+            public class TestMarker {                               
+                @${SpringCoreClasses.BEAN}
+                public B b() { return new B(); }
+            }
+            """.trimIndent()
+
+        myFixture.configureByText("TestMarker.java", text)
+        myFixture.doHighlighting()
+
+        val allBeanGutters = SpringGutterTestUtil.getAllBeanGuttersByIcon(myFixture, SpringIcons.SpringBean)
+        val gutterTargetString = SpringGutterTestUtil.getGutterTargetString(allBeanGutters)
+        assertEquals(1, gutterTargetString.size)
+        assertEquals(listOf("b"), gutterTargetString[0])
+    }
+
+    fun testArrayUsageBaseType() {
+        myFixture.copyFileToProject("BeanUsagesClasses.java")
+        myFixture.addClass(
+            """                                    
+            @${SpringCoreClasses.COMPONENT}
+            class TestUsages {
+                @${SpringCoreClasses.AUTOWIRED}
+                I[] i;                
+            }
+        """.trimIndent()
+        )
+
+        @Language("JAVA") val text = """           
+            @${SpringCoreClasses.CONFIGURATION}
+            public class TestMarker {                               
+                @${SpringCoreClasses.BEAN}
+                public B b() { return new B(); }
+            }
+            """.trimIndent()
+
+        myFixture.configureByText("TestMarker.java", text)
+        myFixture.doHighlighting()
+
+        val allBeanGutters = SpringGutterTestUtil.getAllBeanGuttersByIcon(myFixture, SpringIcons.SpringBean)
+        val gutterTargetString = SpringGutterTestUtil.getGutterTargetString(allBeanGutters)
+        assertEquals(1, gutterTargetString.size)
+        assertEquals(listOf("i"), gutterTargetString[0])
+    }
+
+    //todo error
+    fun testArrayBeanUsage() {
+        myFixture.copyFileToProject("BeanUsagesClasses.java")
+        myFixture.addClass(
+            """                                    
+            @${SpringCoreClasses.COMPONENT}
+            class TestUsages {
+                @${SpringCoreClasses.AUTOWIRED}
+                List<B[]> listsB;               
+            }
+        """.trimIndent()
+        )
+
+        @Language("JAVA") val text = """           
+            @${SpringCoreClasses.CONFIGURATION}
+            public class TestMarker {                               
+                @${SpringCoreClasses.BEAN}
+                B[] carr() { return new B[0]; }
+            }
+            """.trimIndent()
+
+        myFixture.configureByText("TestMarker.java", text)
+        myFixture.doHighlighting()
+
+        val allBeanGutters = SpringGutterTestUtil.getAllBeanGuttersByIcon(myFixture, SpringIcons.SpringBean)
+        val gutterTargetString = SpringGutterTestUtil.getGutterTargetString(allBeanGutters)
+        //assertEquals(1, gutterTargetString.size)
+        //assertEquals(listOf("listsB"), gutterTargetString[0])
+    }
+}